+++ sktime/performance_metrics/forecasting/_medsse.py
@@ -1,0 +1,150 @@
+#!/usr/bin/env python3 -u
+# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
+"""Metrics classes to assess performance on forecasting task.
+
+Classes named as ``*Score`` return a value to maximize: the higher the better.
+Classes named as ``*Error`` or ``*Loss`` return a value to minimize:
+the lower the better.
+"""
+
+from sktime.performance_metrics.forecasting._base import (
+    _ScaledMetricTags,
+    BaseForecastingErrorMetricFunc,
+)
+from sktime.performance_metrics.forecasting._functions import (
+    median_squared_scaled_error,
+)
+
+
+class MedianSquaredScaledError(_ScaledMetricTags, BaseForecastingErrorMetricFunc):
+    """Median squared scaled error (MdSSE) or root median squared scaled error (RMdSSE).
+
+    If ``square_root`` is False then calculates MdSSE, otherwise calculates RMdSSE if
+    ``square_root`` is True. Both MdSSE and RMdSSE output is non-negative floating
+    point. The best value is 0.0.
+
+    This is a squared variant of the MdASE loss metric. Like MASE and other
+    scaled performance metrics this scale-free metric can be used to compare
+    forecast methods on a single series or between series.
+
+    This metric is also suited for intermittent-demand series because it
+    will not give infinite or undefined values unless the training data
+    is a flat timeseries. In this case the function returns a large value
+    instead of inf.
+
+    Works with multioutput (multivariate) timeseries data
+    with homogeneous seasonal periodicity.
+
+    Parameters
+    ----------
+    sp : int, default = 1
+        Seasonal periodicity of data.
+    square_root : bool, default = False
+        Whether to take the square root of the metric
+
+    multioutput : {'raw_values', 'uniform_average'} or array-like of shape \
+            (n_outputs,), default='uniform_average'
+        Defines how to aggregate metric for multivariate (multioutput) data.
+
+        * If array-like, values used as weights to average the errors.
+        * If ``'raw_values'``,
+          returns a full set of errors in case of multioutput input.
+        * If ``'uniform_average'``,
+          errors of all outputs are averaged with uniform weight.
+
+    multilevel : {'raw_values', 'uniform_average', 'uniform_average_time'}
+        Defines how to aggregate metric for hierarchical data (with levels).
+
+        * If ``'uniform_average'`` (default),
+          errors are mean-averaged across levels.
+        * If ``'uniform_average_time'``,
+          metric is applied to all data, ignoring level index.
+        * If ``'raw_values'``,
+          does not average errors across levels, hierarchy is retained.
+
+    by_index : bool, default=False
+        Determines averaging over time points in direct call to metric object.
+
+        * If False, direct call to the metric object averages over time points,
+          equivalent to a call of the``evaluate`` method.
+        * If True, direct call to the metric object evaluates the metric at each
+          time point, equivalent to a call of the ``evaluate_by_index`` method.
+
+    See Also
+    --------
+    MeanAbsoluteScaledError
+    MedianAbsoluteScaledError
+    MedianSquaredScaledError
+
+    References
+    ----------
+    M5 Competition Guidelines.
+
+    https://mofc.unic.ac.cy/wp-content/uploads/2020/03/M5-Competitors-Guide-Final-10-March-2020.docx
+
+    Hyndman, R. J and Koehler, A. B. (2006). "Another look at measures of
+    forecast accuracy", International Journal of Forecasting, Volume 22, Issue 4.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from sktime.performance_metrics.forecasting import MedianSquaredScaledError
+    >>> y_train = np.array([5, 0.5, 4, 6, 3, 5, 2])
+    >>> y_true = np.array([3, -0.5, 2, 7, 2])
+    >>> y_pred = np.array([2.5, 0.0, 2, 8, 1.25])
+    >>> rmdsse = MedianSquaredScaledError(square_root=True)
+    >>> rmdsse(y_true, y_pred, y_train=y_train)
+    np.float64(0.16666666666666666)
+    >>> y_train = np.array([[0.5, 1], [-1, 1], [7, -6]])
+    >>> y_true = np.array([[0.5, 1], [-1, 1], [7, -6]])
+    >>> y_pred = np.array([[0, 2], [-1, 2], [8, -5]])
+    >>> rmdsse(y_true, y_pred, y_train=y_train)
+    np.float64(0.1472819539849714)
+    >>> rmdsse = MedianSquaredScaledError(multioutput='raw_values', square_root=True)
+    >>> rmdsse(y_true, y_pred, y_train=y_train)
+    array([0.08687445, 0.20203051])
+    >>> rmdsse = MedianSquaredScaledError(multioutput=[0.3, 0.7], square_root=True)
+    >>> rmdsse(y_true, y_pred, y_train=y_train)
+    np.float64(0.16914781383660782)
+    """
+
+    func = median_squared_scaled_error
+
+    def __init__(
+        self,
+        multioutput="uniform_average",
+        multilevel="uniform_average",
+        sp=1,
+        square_root=False,
+        by_index=False,
+    ):
+        self.sp = sp
+        self.square_root = square_root
+        super().__init__(
+            multioutput=multioutput,
+            multilevel=multilevel,
+            by_index=by_index,
+        )
+
+    @classmethod
+    def get_test_params(cls, parameter_set="default"):
+        """Return testing parameter settings for the estimator.
+
+        Parameters
+        ----------
+        parameter_set : str, default="default"
+            Name of the set of test parameters to return, for use in tests. If no
+            special parameters are defined for a value, will return ``"default"`` set.
+
+        Returns
+        -------
+        params : dict or list of dict, default = {}
+            Parameters to create testing instances of the class
+            Each dict are parameters to construct an "interesting" test instance, i.e.,
+            ``MyClass(**params)`` or ``MyClass(**params[i])`` creates a valid test
+            instance.
+            ``create_test_instance`` uses the first (or only) dictionary in ``params``
+        """
+        params1 = {}
+        params2 = {"sp": 2, "square_root": True}
+        return [params1, params2]
